requirements.txt
Django==1.11.2
django-debug-toolbar==1.8
django-filter==1.0.4
djangorestframework==3.6.3
